deduplicate the parameter translation machinery
